requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
requests==2.7.0
beautifulsoup4==4.8.2
mysql-connector-python==8.0.19
certifi==2019.11.28
lxml==4.5.0
py-healthcheck==1.9.0
Flask_SocketIO